internal(vx-demo): remove unused codeblocks/, move Codeblock to components/
